requirements.txt
git+https://github.com/craftoid/twitterbot
pillow
numpy
moviepy
imageio